+++ source/ProcessManager.Abstractions/Contracts/EnqueueActorMessagesV1.cs
@@ -13,12 +13,23 @@
 // limitations under the License.
 
 using System.Text.Json;
+using Energinet.DataHub.ProcessManager.Abstractions.Api.Model.OrchestrationDescription;
 
 namespace Energinet.DataHub.ProcessManager.Abstractions.Contracts;
 
 public partial class EnqueueActorMessagesV1
 {
     public const string MajorVersion = nameof(EnqueueActorMessagesV1);
+
+    /// <summary>
+    /// Build the service bus message subject using the provided orchestration unique name.
+    /// </summary>
+    public static string BuildServiceBusMessageSubject(string orchestrationUniqueName) => $"Enqueue_{orchestrationUniqueName.ToLower()}";
+
+    /// <summary>
+    /// Build the service bus message subject using the provided orchestration unique name.
+    /// </summary>
+    public static string BuildServiceBusMessageSubject(OrchestrationDescriptionUniqueNameDto uniqueName) => BuildServiceBusMessageSubject(uniqueName.Name);
 
     public void SetData<TData>(TData data)
         where TData : class