+++ source/ProcessManager.Orchestrations/Processes/BRS_026_028/BRS_028/V1/OrchestrationDescriptionBuilder.cs
@@ -29,7 +29,7 @@ internal class OrchestrationDescriptionBuilder : IOrchestrationDescriptionBuilde
 
         description.ParameterDefinition.SetFromType<RequestCalculatedWholesaleServicesInputV1>();
 
-        description.AppendStepDescription("Asynkron validering");
+        description.AppendStepDescription("Forretningsvalidering");
         description.AppendStepDescription("Udsend beskeder");
 
         return description;