+++ source/Shared/Tests/Fixtures/Extensions/ProcessManagerClientExtensions.cs
@@ -21,6 +21,15 @@ namespace Energinet.DataHub.ProcessManager.Shared.Tests.Fixtures.Extensions;
 
 public static class ProcessManagerClientExtensions
 {
+    private const int TimeLimitInSeconds = 60;
+
+    /// <summary>
+    /// Wait for an orchestration instance, resolving true/false by using the given <paramref name="comparer"/> function.
+    /// <remarks>Returns false if an orchestration instance isn't found in <see cref="TimeLimitInSeconds"/> seconds.</remarks>
+    /// </summary>
+    /// <param name="client"></param>
+    /// <param name="idempotencyKey">The idempotency key of the orchestration instance, used to find the correct orchestration instance.</param>
+    /// <param name="comparer">The comparer method used to determine whether the orchestration instance is in a correct state.</param>
     public static async Task<(bool Succes, OrchestrationInstanceTypedDto<TInput>? OrchestrationInstance)>
         TryWaitForOrchestrationInstance<TInput>(
             this IProcessManagerClient client,
@@ -46,9 +55,53 @@ public static class ProcessManagerClientExtensions
 
                 return comparer(orchestrationInstance);
             },
-            timeLimit: TimeSpan.FromSeconds(60),
+            timeLimit: TimeSpan.FromSeconds(TimeLimitInSeconds),
             delay: TimeSpan.FromSeconds(1));
 
         return (success, orchestrationInstance);
+    }
+
+    /// <summary>
+    /// Wait for an orchestration instance to be terminated with the given <paramref name="terminationState"/>.
+    /// <remarks>Returns false if not resolved in <see cref="TimeLimitInSeconds"/> seconds.</remarks>
+    /// </summary>
+    /// <param name="client"></param>
+    /// <param name="idempotencyKey">The idempotency key of the orchestration instance, used to find the correct orchestration instance.</param>
+    /// <param name="terminationState">The termination state the orchestration instance should be in.</param>
+    public static Task<(bool Succes, OrchestrationInstanceTypedDto<TInput>? OrchestrationInstance)> WaitForOrchestrationInstanceTerminated<TInput>(
+        this IProcessManagerClient client,
+        string idempotencyKey,
+        OrchestrationInstanceTerminationState terminationState)
+        where TInput : class, IInputParameterDto
+    {
+        return client.TryWaitForOrchestrationInstance<TInput>(
+            idempotencyKey,
+            (orchestrationInstance) =>
+                orchestrationInstance.Lifecycle.State == OrchestrationInstanceLifecycleState.Terminated
+                && orchestrationInstance.Lifecycle.TerminationState == terminationState);
+    }
+
+    /// <summary>
+    /// Wait for an orchestration instance step to be in the <see cref="StepInstanceLifecycleState.Running"/> state.
+    /// <remarks>Returns false if not resolved in <see cref="TimeLimitInSeconds"/> seconds.</remarks>
+    /// </summary>
+    /// <param name="client"></param>
+    /// <param name="idempotencyKey">The idempotency key of the orchestration instance, used to find the correct orchestration instance.</param>
+    /// <param name="stepSequence">The sequence number of the step that should be in the <see cref="StepInstanceLifecycleState.Running"/> state.</param>
+    public static Task<(bool Succes, OrchestrationInstanceTypedDto<TInput>? OrchestrationInstance)> WaitForStepToBeRunning<TInput>(
+        this IProcessManagerClient client,
+        string idempotencyKey,
+        int stepSequence)
+            where TInput : class, IInputParameterDto
+    {
+        return client.TryWaitForOrchestrationInstance<TInput>(
+            idempotencyKey,
+            (orchestrationInstance) =>
+            {
+                var enqueueActorMessagesStep = orchestrationInstance.Steps
+                    .Single(s => s.Sequence == stepSequence);
+
+                return enqueueActorMessagesStep.Lifecycle.State == StepInstanceLifecycleState.Running;
+            });
     }
 }